requirements.txt
numpy
chromadb
psycopg2-binary
psutil
matplotlib
pandas
